+++ packages/dates/src/constants.ts
@@ -1,0 +1,1 @@
+export const NOW = new Date();
