+++ packages/dates/src/utils/index.ts
@@ -1,0 +1,7 @@
+export { default as isOnOrAfterToday } from './isOnOrAfterToday';
+export { default as isOnOrBeforeToday } from './isOnOrBeforeToday';
+export { default as isOnOrAfterDate } from './isOnOrAfterDate';
+export { default as isOnOrBeforeDate } from './isOnOrBeforeDate';
+
+export * from './types';
+export * from './misc';
