+++ packages/utils/src/date/index.ts
@@ -1,10 +1,6 @@
 export * from './addSub';
 export { default as areEqual } from './areEqual';
 export { default as diff } from './diff';
-export { default as isOnOrAfterToday } from './isOnOrAfterToday';
-export { default as isOnOrBeforeToday } from './isOnOrBeforeToday';
-export { default as isOnOrAfterDate } from './isOnOrAfterDate';
-export { default as isOnOrBeforeDate } from './isOnOrBeforeDate';
 export { default as sort } from './sort';
 export { default as switchTenseForDate } from './switchTenseForDate';
 export * from './misc';
